Review/requirements.txt
pydantic==2.9.2
scikit-learn==1.5.2
numpy==2.1.3
groq==1.0.0
streamlit==1.40.0
python-dotenv==1.0.1

